Move interfaces to core
